# docs/api.yaml
openapi: 3.0.1
info:
  title: NAPA project
  description: NAPA trainee assignment
  version: "1"
...
paths:
...
  /countries:
    post:
      tags: ["countries"]
      description: Adds a new country to the data base
      operationId: addCountry
      requestBody:
       description: Country details
       required: true
       content:
        application/json:
          schema:
            type: object
            description: The name of the country and if it has been visited
            properties:
              name:
                $ref: '#/components/schemas/CountryName'
              visited:
                $ref: '#/components/schemas/CountryVisited'
      responses:
        '201':
          description: Country created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipName'
        '400':
          description: Invalid input or parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
    get:
      tags: ["countries"]
      description: Retrieves the names of all the countries in the data base
      operationId: getCountries
      responses:
        '200':
          description: The list of all country names.
          content:
            application/json:
              schema:
                description: Country names array
                type: array
                items:
                  $ref: '#/components/schemas/CountryName'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /countries/{name}:
    parameters:
     - name: name
       in: path
       required: true  
       description: The name for a country.
       schema:
        $ref: '#/components/schemas/CountryName'
    put:
      tags: ["countries"]
      description: Update a countries name
      operationId: getCountryName
      requestBody:
       description: New country name
       required: true
       content:
        application/json:
          schema:
           $ref: '#/components/schemas/CountryName'
      responses:
        '204':
          description: Successful update
        '403':
         description: Forbidden. Country name already exists
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/ErrorMessage'
        '400':
          description: Invalid input or parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: Country not found.
          content:
           application/json:
             schema:
               $ref: '#/components/schemas/ErrorMessage'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
    get:
      tags: ["countries"]
      description: See if a country has been visited
      operationId: getCountryVisited
      responses:
       '200':
         description: If a country has been visited
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/CountryVisited'
       '404':
          description: Country not found.
          content:
           application/json:
             schema:
               $ref: '#/components/schemas/ErrorMessage'
       '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
    delete:
      tags: ["countries"]
      description: Delete a country
      operationId: deleteCountry
      responses:
        '204':
          description: Successful delete
        '404':
          description: Country not found.
          content:
           application/json:
             schema:
               $ref: '#/components/schemas/ErrorMessage'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'  
...
components: 
  schemas:
...
    ShipName:
      type: string
      description: The name of a ship.  Also functions as unique identifier.
      pattern: '^.*?$'
      example: Titanic
    CountryName:
      type: string
      description: The name of a country. Also functions as unique identifier.
      pattern: '^.*?$'
      example: France
    CountryVisited:
      type: boolean
      description: Indicates if a country has been visited.
...
    ErrorMessage:
      type: string
      pattern: '^.*?$'
      example: Error
      description: Error Message